Move common utilities into utils module
