+++ rq_dashboard/utils.py
@@ -1,0 +1,68 @@
+from functools import wraps
+from math import ceil
+
+import arrow
+from flask import current_app, url_for
+
+
+def jsonify(f):
+    @wraps(f)
+    def _wrapped(*args, **kwargs):
+        from flask import jsonify as flask_jsonify
+        try:
+            result_dict = f(*args, **kwargs)
+        except Exception as e:
+            result_dict = dict(status='error')
+            if current_app.config['DEBUG']:
+                result_dict['reason'] = str(e)
+                from traceback import format_exc
+                result_dict['exc_info'] = format_exc()
+        return flask_jsonify(**result_dict)
+    return _wrapped
+
+
+def pagination_window(total_items, cur_page, per_page=5, window_size=10):
+    all_pages = range(1, int(ceil(total_items / float(per_page))) + 1)
+    result = all_pages
+    if window_size >= 1:
+        temp = min(
+            len(all_pages) - window_size,
+            (cur_page - 1) - int(ceil(window_size / 2.0))
+        )
+        pages_window_start = max(0, temp)
+        pages_window_end = pages_window_start + window_size
+        result = all_pages[pages_window_start:pages_window_end]
+    return result
+
+
+def remove_none_values(input_dict):
+    return {k: v for (k, v) in input_dict.items() if v is not None}
+
+
+def serialize_date(dt):
+    if dt is None:
+        return None
+    return arrow.get(dt).to('UTC').datetime.isoformat()
+
+
+def serialize_queues(queues):
+    return [
+        dict(
+            name=q.name,
+            count=q.count,
+            url=url_for('.overview', queue_name=q.name))
+        for q in queues
+    ]
+
+
+def serialize_job(job, scheduled_for=None):
+    return dict(
+        id=job.id,
+        created_at=serialize_date(job.created_at),
+        enqueued_at=serialize_date(job.enqueued_at),
+        scheduled_for=serialize_date(scheduled_for) if scheduled_for else None,
+        ended_at=serialize_date(job.ended_at),
+        origin=job.origin,
+        result=job._result,
+        exc_info=job.exc_info,
+        description=job.description)
